agentic-document-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

numpy==1.26.4
opencv-python==4.6.0.66
pillow==10.0.1

paddleocr==2.7.3
paddlepaddle==2.6.2

transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.206

python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2
pdf2image
